+++ src/assets/config/models.ts
@@ -3,13 +3,11 @@ import { Models } from '../../app/types/model.types'
 // provider = name of model creator
 // name = display name of the model
 // id = the API id or name of the model
-// inputPrice = $/1M tokens in prompt
-// outputPrice = $/1M tokens in model output
+// inputPrice = Either a flat price ($/1M tokens in prompt) or a tiered
+//              price depending on token count.
+// outputPrice = Either a flat price ($/1M tokens in model output) or a
+//               tiered price depending on token count.
 // rpm = max requests per minute the model accepts at current usage tier
-// reasoning = (optional, required for reasoning models) reasoning effort
-//             constraint for reasoning models, supported values are
-//             `none`, `minimal`, `low`, `medium` and `high` depending
-//             on the model
 // default = (optional) boolean indicating which model is the default
 // parameters = an object with model parameters:
 //     maxImageShortsidePx = (optional) max supported image short side